Move SettingsModal into mafia
