+++ ubg/package.json
@@ -15,7 +15,6 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-elastic-carousel": "^0.6.7",
-    "react-hook-form": "^6.3.0",
     "react-player": "^2.6.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
