+++ ubg/src/common/WaitingRoom.js
@@ -3,7 +3,7 @@ import {makeStyles} from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 import Modal from '@material-ui/core/Modal';
-import SettingsModal from './SettingsModal';
+import SettingsModal from '../mafia/SettingsModal';
 import PropTypes from 'prop-types';
 
 const useStyles = makeStyles((theme) => ({
